HREPTWO-4823: use namespace as old & new prefix when cnd is reloaded
When the CND is reloaded, i.e node types are re-initialized but the namespace version is not bumped, no new prefix is registered.  Nodes that are visited will receive primary and mixin types that do not exist (newPrefix + ':' + localName).  Use the invariant prefix as both the old and the new prefix.