REPO-240: use copyTo step directly
The DefaultWorkflow used the FolderWorkflow#copy method, which uses a few
indirections before copying the nodes.  These workflow indirections did not
have the "translation" context and so the information that the translation ids
should be retained, was lost.

The DefaultWorkflow now uses the same copy mechanism that is used by the
reviewed actions workflow, based on the EmbedWorkflow.