REPO-754: consolidate node copying
Consolidate node copying code.  Copying between nodes, from the version
history, with variables, etcetera.

The folder workflow impl #expand method has been removed; afaics, it expanded
values ${variable} to the property on the source node with name 'variable'.  I
did not see it being used.

The goal of the consolidation is to reduce the number of errors that occur when
implementing copy operations.