REPO-811: drop usage of hippo:harddocument, move hippo:paths to hippo:document
The use of harddocument has been replaced with mix:versionable throughout the
repository.  This mixin can very likely be dropped in most of these cases.

The HippoNodeTypeRegistry skips checks for system namespaces, assuming that we
know what we're doing.  The PathMigrationTest verifies that changing the types
is safe to do in a running system.  (existing sessions will continue to
function)

The softdocument mixin is now set on all virtualized hippo:document instances.