REPO-886 Fix array vulnerability
Exposing array fields allows users to mutate array values.
This should not be allowed for the static supported resolutions field.
I fixed it by making the static fields an immutable list and using that
to return an new array instance on each invocation. A new method has 
been introduced that returns an iterable and the old method has bee deprecated.
In addition I removed the logger, because it was not used.