CMS7-7572 Fix visibility of activity stream
The workflow event logger will set the system flag in the event log to true
if the event was triggered by a system user, so that the value of this flag 
will be persisted to the repository. 
I modified the BroadCastThread so that it will read all the logInfo node
properties and set them on the WorkFlowEvent by calling event.set(key, value).
This way all available properties will always be set, even if new properties
will be added later. To simplify getting the values of properties I have added
the ValueGetter interface and the accompanying impls.