REPO-1151 support for path based facet rules in domains : If a facetrule has hipposys:facet = jcr:path and hipposys:type = Reference, then that facetrule is a match for any node that is equal to or a descendant of hipposys:value, for example hipposys:value = /content/documents/myhippoproject/news
Obviously, this facetrule which can be AND-ed with other facet rules to downscope read access below, say /content/documents/myhippoproject/news. An example of such downscope of read access is shown in the unit test by user 'alice'. I will add another test scenario for a domain that has a facetrule matching /x/y (and descendants) that is AND-ed with a facetrule that disallowes (unmatches) /x/y/z