REPO-1179: wait for 10 seconds instead of 5
In the case that the job runs just before the event is logged, it will take (at least) 5 seconds before the next run occurs.  This leaves a 100ms window for the broadcast thread to wake up and run the job that should notify the listener.
Such a window may be too short on a heavily loaded system.