REPO-1497 Fix parent/child axis queries
In FacetedAuthorizationTest I added a lot of different parent/child/deref
queries that result in hierarchical queries for which we need to include
the lucene filter in the backing queries. These queries validate the changes
I made to jackrabbit in REPO-1503

In ServicingSearchIndex I added a performance impr by checking whether the
query is a match all docs query: If it is, it is more efficient to return
a null filter

In HippoLuceneQueryHits I added a null check because the filter can (and
even already could before) be null

In HippoIndexSearcher I pass the authorizationFilter to the JackrabbitIndexSearcher for which
I changed the JackrabbitIndexSearcher constructor in REPO-1503
I removed the 'if (query instanceof JackrabbitQuery) {'
check : That logic did not include our authorization filter! Also our 'sort'
is never empty afaics (we have an empty sort object, not null) which in the
end resulted that ((JackrabbitQuery) query).execute(this, getSession(), sort);
returned null any way.