REPO-1545: drop unnecessary pre-emptive webfiles bundle delete before reload
The PostStartupTask will invoke the WebFileService to import the new bundle which also,
and without intermediate save, first removes the existing webfiles resources.
The pre-emptive delete in WebFileBundleInstruction is therefore not needed, while it can cause
temporarily unavailability of the existing webfiles on other active (clustered) repository instances
because the initialization processor will perform an intermediate session save.