REPO-1636:Take into account daylight saving
* Take into account time zones to calculate the expected length of the
  interval between dates.

java.util.Date has not been designed for internationalization and does
not have a timezone concept. In the unit test java.util.Date was used to
calcutate the length of the interval in days. Since the timezone was not
taken into account that fails if the moment of changing to daylight
saving and back lies within the interval.

The tested method returns the correct value.