REPO-1673 [Backport 3.2.8] make sure to synchronize on the jcr session to avoid thread safety issues
If we synchronize on the monitor object, there is room for concurrent access on the jcr session
because code that created the lock can call for example #isLive which can work on the
same jcr session as the background executor job does that refreshes the lock.

Also made sure that if #setTimeout fails with a repository exception, that this exception
is propagated and results in a lock exception. Before this exception would only be logged,
and situations were seen that the background thread that refreshes the lock would log
a repository exception in #setTimeout every 52 seconds (default 1 minute minus 8 second delay)
were it should abandon the refreshing completely.

In case of an exception, always log the stacktrace and not just in debug mode

(cherry picked from commit 2aa0536f7e668bbe50007180f4091eafaf6a8175)