REPO-1705 remove hippostd:stateSummary properties
These projects do not contain tests that rely on the deleted node.
The underlying problem is that hippostd:stateSummary properties are not converted as well as ignored by all HCM processing. For most modules that is not a problem, as they also load the configuration of the workflow module which contains the configuration for the derived data engine which calculates this correctly.