REPO-1704 move bootstrap of hippostd-queries from repository-config to repository-workflow module to properly and always resolve hippostd:stateSummary dynamically
The hippostd-queries configuration uses and depends on the derived data engine to 'derive' the mandatory hippostd:stateSummary for the new-document template (primaryType: hippostd:document)
But the derived data engine *configuration* needed to do this is bootstrapped by the repository-workflow, which comes *after* repository-config dependency wise.
Moving the hippstd-queries bootstrap to repository-workflow therefore is needed to *not* needing to bootstrap a dummy/fake value for hippostd:stateSummary (which HCM also no longer can/will do).

Any *test* usage of hippostd-queries should already also depend on hippo-workflow (as it really only becomes useful within context of workflow), so this move should not cause build any problems,
or otherwise such test usage should then also (or instead) test depend on repository-workflow to fix it.