REPO-1859 Make documentworkflow aware of context outside repository
Make an initial payload accessible to the DocumentWorkflow. This makes
it possible to include information that is available to the http
session.

Introduce DocumentWorkflowTransition with builder to facilitate
various combination of parameters that have to be passed on the
triggerAction method of the ScxmlExecutor.

Reordered access modifiers according the java specs.