REPO-1835 Refactor DocumentWorkflowImpl
Goal is to be able to set an initializationPayload via the
DocumentWorkflow interface to all actions and hints.

Introduce DocumentWorkflowTransition with builder to facilitate
various combination of parameters that have to be passed on the
triggerAction method of the ScxmlExecutor.

Do not refactor the request... methods for now. Refactoring would need
an interface for Request or more state in the DocumentWorkflowTransition
and subsequent logic in the transition method.

Add Enum for action to avoid typos in calls to api.